+++ src/Exception/CommonProblemDetailsExceptionTrait.php
@@ -1,11 +1,15 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-problem-details for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-problem-details/blob/master/LICENSE.md New BSD License
  */
 
+declare(strict_types=1);
+
 namespace Zend\ProblemDetails\Exception;
+
+use function array_merge;
 
 /**
  * Common functionality for ProblemDetailsExceptionInterface implementations.