+++ src/ProblemDetailsNotFoundHandlerFactory.php
@@ -1,9 +1,11 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-problem-details for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-problem-details/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace Zend\ProblemDetails;
 
@@ -13,8 +15,6 @@ class ProblemDetailsNotFoundHandlerFactory
 {
     public function __invoke(ContainerInterface $container) : ProblemDetailsNotFoundHandler
     {
-        return $container->has(ProblemDetailsResponseFactory::class)
-            ? new ProblemDetailsNotFoundHandler($container->get(ProblemDetailsResponseFactory::class))
-            : new ProblemDetailsNotFoundHandler();
+        return new ProblemDetailsNotFoundHandler($container->get(ProblemDetailsResponseFactory::class));
     }
 }