+++ src/ProblemDetailsResponseFactory.php
@@ -1,9 +1,11 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-problem-details for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-problem-details/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace Zend\ProblemDetails;
 
@@ -12,11 +14,27 @@ use Fig\Http\Message\StatusCodeInterface as StatusCode;
 use Negotiation\Negotiator;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Psr\Http\Message\StreamInterface;
 use Spatie\ArrayToXml\ArrayToXml;
 use Throwable;
-use Zend\Diactoros\Response;
-use Zend\Diactoros\Stream;
+
+use function array_merge;
+use function array_walk_recursive;
+use function get_class;
+use function get_resource_type;
+use function is_array;
+use function is_int;
+use function is_resource;
+use function json_decode;
+use function json_encode;
+use function preg_replace;
+use function print_r;
+use function sprintf;
+use function strpos;
+
+use const JSON_PRESERVE_ZERO_FRACTION;
+use const JSON_PRETTY_PRINT;
+use const JSON_UNESCAPED_SLASHES;
+use const JSON_UNESCAPED_UNICODE;
 
 /**
  * Create a Problem Details response.
@@ -133,18 +151,6 @@ class ProblemDetailsResponseFactory
     ];
 
     /**
-     * Factory for generating an empty response body.
-     *
-     * If none is provided, defaults to a closure that returns an empty
-     * zend-diactoros Stream instance using a php://temp stream.
-     *
-     * The factory MUST return a StreamInterface
-     *
-     * @var callable
-     */
-    private $bodyFactory;
-
-    /**
      * Whether or not to include debug details.
      *
      * Debug details are only included for responses created from throwables,
@@ -158,20 +164,22 @@ class ProblemDetailsResponseFactory
     /**
      * JSON flags to use when generating JSON response payload.
      *
-     * Defaults to JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE | JSON_PRESERVE_ZERO_FRACTION
+     * On non-debug mode:
+     * defaults to JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE | JSON_PRESERVE_ZERO_FRACTION
+     * On debug mode:
+     * defaults to JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE | JSON_PRESERVE_ZERO_FRACTION
      *
      * @var int
      */
     private $jsonFlags;
 
     /**
-     * Response prototype to use when generating Problem Details responses.
+     * Factory to use to generate prototype response used when generating a
+     * problem details response.
      *
-     * Defaults to a zend-diactoros response if none is injected.
-     *
-     * @var ResponseInterface
+     * @var callable
      */
-    private $response;
+    private $responseFactory;
 
     /**
      * Flag to enable show exception details in detail field.
@@ -193,18 +201,24 @@ class ProblemDetailsResponseFactory
     private $defaultDetailMessage;
 
     public function __construct(
+        callable $responseFactory,
         bool $isDebug = self::EXCLUDE_THROWABLE_DETAILS,
         int $jsonFlags = null,
-        ResponseInterface $response = null,
-        callable $bodyFactory = null,
         bool $exceptionDetailsInResponse = false,
         string $defaultDetailMessage = self::DEFAULT_DETAIL_MESSAGE
     ) {
+        // Ensures type safety of the composed factory
+        $this->responseFactory = function () use ($responseFactory) : ResponseInterface {
+            return $responseFactory();
+        };
         $this->isDebug = $isDebug;
-        $this->jsonFlags = $jsonFlags
-            ?: JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE | JSON_PRESERVE_ZERO_FRACTION;
-        $this->response = $response ?: new Response();
-        $this->bodyFactory = $bodyFactory ?: Closure::fromCallable([$this, 'generateStream']);
+        if (! $jsonFlags) {
+            $jsonFlags = JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE | JSON_PRESERVE_ZERO_FRACTION;
+            if ($isDebug) {
+                $jsonFlags = JSON_PRETTY_PRINT | $jsonFlags;
+            }
+        }
+        $this->jsonFlags = $jsonFlags;
         $this->exceptionDetailsInResponse = $exceptionDetailsInResponse;
         $this->defaultDetailMessage = $defaultDetailMessage;
     }
@@ -334,30 +348,14 @@ class ProblemDetailsResponseFactory
         );
     }
 
-    /**
-     * @throws Exception\InvalidResponseBodyException
-     */
     protected function generateResponse(int $status, string $contentType, string $payload) : ResponseInterface
     {
-        $body = ($this->bodyFactory)();
-        if (! $body instanceof StreamInterface) {
-            throw new Exception\InvalidResponseBodyException(sprintf(
-                'The factory for generating a problem details response body stream did not return a %s',
-                StreamInterface::class
-            ));
-        }
+        $response = ($this->responseFactory)();
+        $response->getBody()->write($payload);
 
-        $body->write($payload);
-
-        return $this->response
+        return $response
             ->withStatus($status)
-            ->withHeader('Content-Type', $contentType)
-            ->withBody($body);
-    }
-
-    private function generateStream() : StreamInterface
-    {
-        return new Stream('php://temp', 'wb+');
+            ->withHeader('Content-Type', $contentType);
     }
 
     private function getResponseGenerator(ServerRequestInterface $request) : callable