+++ src/ProblemDetailsResponseFactoryFactory.php
@@ -1,9 +1,11 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-problem-details for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-problem-details/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace Zend\ProblemDetails;
 
@@ -20,31 +22,11 @@ class ProblemDetailsResponseFactoryFactory
         $problemDetailsConfig = $config['problem-details'] ?? [];
         $jsonFlags = $problemDetailsConfig['json_flags'] ?? null;
 
-        $responsePrototype = $this->getResponsePrototype($container);
-
-        $streamFactory = $container->has('Zend\ProblemDetails\StreamFactory')
-            ? $container->get('Zend\ProblemDetails\StreamFactory')
-            : null;
-
         return new ProblemDetailsResponseFactory(
+            $container->get(ResponseInterface::class),
             $includeThrowableDetail,
             $jsonFlags,
-            $responsePrototype,
-            $streamFactory,
             $includeThrowableDetail
         );
-    }
-
-    /**
-     * @return null|ResponseInterface
-     */
-    private function getResponsePrototype(ContainerInterface $container)
-    {
-        if (! $container->has(ResponseInterface::class)) {
-            return null;
-        }
-
-        $response = $container->get(ResponseInterface::class);
-        return is_callable($response) ? $response() : $response;
     }
 }