+++ test/Exception/ProblemDetailsExceptionInterfaceTest.php
@@ -1,15 +1,20 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-problem-details for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-problem-details/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace ZendTest\ProblemDetails\Exception;
 
 use PHPUnit\Framework\TestCase;
 use Zend\ProblemDetails\Exception\CommonProblemDetailsExceptionTrait;
 use Zend\ProblemDetails\Exception\ProblemDetailsExceptionInterface;
+
+use function json_decode;
+use function json_encode;
 
 class ProblemDetailsExceptionTest extends TestCase
 {