+++ test/ProblemDetailsResponseFactoryFactoryTest.php
@@ -1,19 +1,30 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-problem-details for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-problem-details/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace ZendTest\ProblemDetails;
 
 use Closure;
+use PHPUnit\Framework\Assert;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ResponseInterface;
-use Zend\Diactoros\Response;
+use ReflectionProperty;
+use RuntimeException;
+use stdClass;
+use TypeError;
 use Zend\ProblemDetails\ProblemDetailsResponseFactory;
 use Zend\ProblemDetails\ProblemDetailsResponseFactoryFactory;
+
+use const JSON_PRESERVE_ZERO_FRACTION;
+use const JSON_PRETTY_PRINT;
+use const JSON_UNESCAPED_SLASHES;
+use const JSON_UNESCAPED_UNICODE;
 
 class ProblemDetailsResponseFactoryFactoryTest extends TestCase
 {
@@ -22,11 +33,48 @@ class ProblemDetailsResponseFactoryFactoryTest extends TestCase
         $this->container = $this->prophesize(ContainerInterface::class);
     }
 
-    public function testLackOfOptionalServicesResultsInFactoryUsingDefaults() : void
+    public function assertResponseFactoryReturns(ResponseInterface $expected, ProblemDetailsResponseFactory $factory)
+    {
+        $r = new ReflectionProperty($factory, 'responseFactory');
+        $r->setAccessible(true);
+        $responseFactory = $r->getValue($factory);
+
+        Assert::assertSame($expected, $responseFactory());
+    }
+
+    public function testLackOfResponseServiceResultsInException()
+    {
+        $factory = new ProblemDetailsResponseFactoryFactory();
+        $e = new RuntimeException();
+
+        $this->container->has('config')->willReturn(false);
+        $this->container->get('config')->shouldNotBeCalled();
+        $this->container->get(ResponseInterface::class)->willThrow($e);
+
+        $this->expectException(RuntimeException::class);
+        $factory($this->container->reveal());
+    }
+
+    public function testNonCallableResponseServiceResultsInException()
+    {
+        $factory = new ProblemDetailsResponseFactoryFactory();
+
+        $this->container->has('config')->willReturn(false);
+        $this->container->get('config')->shouldNotBeCalled();
+        $this->container->get(ResponseInterface::class)->willReturn(new stdClass);
+
+        $this->expectException(TypeError::class);
+        $factory($this->container->reveal());
+    }
+
+    public function testLackOfConfigServiceResultsInFactoryUsingDefaults() : void
     {
         $this->container->has('config')->willReturn(false);
-        $this->container->has(ResponseInterface::class)->willReturn(false);
-        $this->container->has('Zend\ProblemDetails\StreamFactory')->willReturn(false);
+
+        $response = $this->prophesize(ResponseInterface::class)->reveal();
+        $this->container->get(ResponseInterface::class)->willReturn(function () use ($response) {
+            return $response;
+        });
 
         $factoryFactory = new ProblemDetailsResponseFactoryFactory();
         $factory = $factoryFactory($this->container->reveal());
@@ -34,13 +82,32 @@ class ProblemDetailsResponseFactoryFactoryTest extends TestCase
         $this->assertInstanceOf(ProblemDetailsResponseFactory::class, $factory);
         $this->assertAttributeSame(ProblemDetailsResponseFactory::EXCLUDE_THROWABLE_DETAILS, 'isDebug', $factory);
         $this->assertAttributeSame(
-            JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE | JSON_PRESERVE_ZERO_FRACTION,
+            JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE | JSON_PRESERVE_ZERO_FRACTION,
             'jsonFlags',
             $factory
         );
 
-        $this->assertAttributeInstanceOf(Response::class, 'response', $factory);
-        $this->assertAttributeInstanceOf(Closure::class, 'bodyFactory', $factory);
+        $this->assertAttributeInstanceOf(Closure::class, 'responseFactory', $factory);
+        $this->assertResponseFactoryReturns($response, $factory);
+    }
+
+    public function testUsesPrettyPrintFlagOnEnabledDebugMode() : void
+    {
+        $this->container->has('config')->willReturn(true);
+        $this->container->get('config')->willReturn([
+            'debug' => true,
+        ]);
+        $this->container->get(ResponseInterface::class)->willReturn(function () {
+        });
+
+        $factoryFactory = new ProblemDetailsResponseFactoryFactory();
+        $factory = $factoryFactory($this->container->reveal());
+
+        $this->assertAttributeSame(
+            JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE | JSON_PRESERVE_ZERO_FRACTION,
+            'jsonFlags',
+            $factory
+        );
     }
 
     public function testUsesDebugSettingFromConfigWhenPresent() : void
@@ -48,8 +115,8 @@ class ProblemDetailsResponseFactoryFactoryTest extends TestCase
         $this->container->has('config')->willReturn(true);
         $this->container->get('config')->willReturn(['debug' => true]);
 
-        $this->container->has(ResponseInterface::class)->willReturn(false);
-        $this->container->has('Zend\ProblemDetails\StreamFactory')->willReturn(false);
+        $this->container->get(ResponseInterface::class)->willReturn(function () {
+        });
 
         $factoryFactory = new ProblemDetailsResponseFactoryFactory();
         $factory = $factoryFactory($this->container->reveal());
@@ -64,66 +131,13 @@ class ProblemDetailsResponseFactoryFactoryTest extends TestCase
         $this->container->has('config')->willReturn(true);
         $this->container->get('config')->willReturn(['problem-details' => ['json_flags' => JSON_PRETTY_PRINT]]);
 
-        $this->container->has(ResponseInterface::class)->willReturn(false);
-        $this->container->has('Zend\ProblemDetails\StreamFactory')->willReturn(false);
+        $this->container->get(ResponseInterface::class)->willReturn(function () {
+        });
 
         $factoryFactory = new ProblemDetailsResponseFactoryFactory();
         $factory = $factoryFactory($this->container->reveal());
 
         $this->assertInstanceOf(ProblemDetailsResponseFactory::class, $factory);
         $this->assertAttributeSame(JSON_PRETTY_PRINT, 'jsonFlags', $factory);
-    }
-
-    public function testUsesResponseServiceFromContainerWhenPresent() : void
-    {
-        $response = $this->prophesize(ResponseInterface::class)->reveal();
-
-        $this->container->has('config')->willReturn(false);
-        $this->container->has(ResponseInterface::class)->willReturn(true);
-        $this->container->get(ResponseInterface::class)->willReturn($response);
-        $this->container->has('Zend\ProblemDetails\StreamFactory')->willReturn(false);
-
-        $factoryFactory = new ProblemDetailsResponseFactoryFactory();
-        $factory = $factoryFactory($this->container->reveal());
-
-        $this->assertInstanceOf(ProblemDetailsResponseFactory::class, $factory);
-        $this->assertAttributeSame($response, 'response', $factory);
-    }
-
-    public function testUsesResponseServiceAsFactoryFromContainerWhenPresent() : void
-    {
-        $response = $this->prophesize(ResponseInterface::class)->reveal();
-        $responseFactory = function () use ($response) {
-            return $response;
-        };
-
-        $this->container->has('config')->willReturn(false);
-        $this->container->has(ResponseInterface::class)->willReturn(true);
-        $this->container->get(ResponseInterface::class)->willReturn($responseFactory);
-        $this->container->has('Zend\ProblemDetails\StreamFactory')->willReturn(false);
-
-        $factoryFactory = new ProblemDetailsResponseFactoryFactory();
-        $factory = $factoryFactory($this->container->reveal());
-
-        $this->assertInstanceOf(ProblemDetailsResponseFactory::class, $factory);
-        $this->assertAttributeSame($response, 'response', $factory);
-    }
-
-    public function testUsesStreamFactoryServiceFromContainerWhenPresent() : void
-    {
-        // @codingStandardsIgnoreStart
-        $streamFactory = function () { };
-        // @codingStandardsIgnoreEnd
-
-        $this->container->has('config')->willReturn(false);
-        $this->container->has(ResponseInterface::class)->willReturn(false);
-        $this->container->has('Zend\ProblemDetails\StreamFactory')->willReturn(true);
-        $this->container->get('Zend\ProblemDetails\StreamFactory')->willReturn($streamFactory);
-
-        $factoryFactory = new ProblemDetailsResponseFactoryFactory();
-        $factory = $factoryFactory($this->container->reveal());
-
-        $this->assertInstanceOf(ProblemDetailsResponseFactory::class, $factory);
-        $this->assertAttributeSame($streamFactory, 'bodyFactory', $factory);
     }
 }